requirements.txt
# requirements.txt
# Python Dependencies for Cricbuzz LiveStats Dashboard
# Student Assignment - Complete Implementation

# Core Framework
streamlit==1.28.0

# Data Processing and Analysis
pandas>=2.2.0
numpy>=1.26.0

# Database Operations
sqlite3
# Note: sqlite3 comes built-in with Python, no separate installation needed

# API Integration
requests==2.31.0

# Data Visualization
plotly==5.15.0

# Environment Configuration
python-dotenv==1.0.0

# Optional: Enhanced Data Processing (if needed)
# openpyxl==3.1.2          # For Excel file support
# xlsxwriter==3.1.9        # For Excel export functionality
# pytz==2023.3             # For timezone handling
# dateutils==0.6.12        # For advanced date operations

# Development Dependencies (Optional)
# pytest==7.4.0            # For unit testing
# black==23.7.0            # For code formatting  
# flake8==6.0.0            # For code linting
# mypy==1.5.1              # For type checking

# Production Dependencies (Optional)
# gunicorn==21.2.0         # For production deployment
# psycopg2-binary==2.9.7   # For PostgreSQL support in production
# redis==4.6.0             # For caching in production


# Note: Install with: pip install -r requirements.txt

